# switch to new op for KeyedTensor.regroup